requirements.txt
torch>=1.7.0
torchvision>=0.8.0
timm>=0.4.5
numpy>=1.19.5
Pillow>=8.0.0
